restructure problems components, shorten RankTable export
