+++ src/features/problems/components/InfoTable/Score.tsx
@@ -3,7 +3,7 @@ import { cookies } from "next/headers";
 import { Score } from "@/components";
 import { getMemoUser } from "@/features/auth";
 
-import { getProblemScore } from "../../../queries";
+import { getProblemScore } from "../../queries";
 
 export const InfoTableScore = async ({ problemId }: { problemId: string }) => {
   const cookiesStore = cookies();
