--- src/features/problems/components/ProblemPage/index.ts
@@ -1,3 +1,0 @@
-export * from "./Editor";
-export * from "./InfoTable";
-export * from "./RecalcScore";
