+++ src/features/problems/components/index.ts
@@ -1,1 +1,3 @@
-export * from "./ProblemPage";
+export * from "./Editor";
+export * from "./InfoTable";
+export * from "./RecalcScore";
